adjusted make files and setup
the libc makefile needs work but im working on it, im also about 104% its gonna be fucked up so heres my safty push
