+++ kernel/src/kernel.h
@@ -1,5 +1,10 @@
 #ifndef _KERNEL_H
 #define _KERNEL_H 1
 
+#define BYTE_TO_KB 1024
+#define BYTE_TO_MB 1048576
+#define BYTE_TO_GB 1073741824
+
+static void halt(void);
 
 #endif
