+++ libc/GNUmakefile
@@ -102,7 +102,7 @@ src/limine.h:
 # See https://sourceware.org/bugzilla/show_bug.cgi?id=31795 for more information.
 bin/$(KERNEL): GNUmakefile linker.ld $(OBJ)
 	mkdir -p "$$(dirname $@)"
-	$(KLD) $(OBJ) ../libc/libc.a $(KLDFLAGS) -o $@
+	$(KLD) $(OBJ) $(KLDFLAGS) -o $@
 	printf '\003' | dd of=$@ bs=1 count=1 seek=16 conv=notrunc 2>/dev/null
 
 # Include header dependencies.
