+++ libc/include/stdlib.h
@@ -1,0 +1,5 @@
+#ifndef _STDLIB_H 
+#define _STDLIB_H 1
+
+
+#endif
