+++ libc/include/string.h
@@ -1,8 +1,10 @@
-#ifndef _MEMORY_H
-#define _MEMORY_H 1
+#ifndef _STRING_H 
+#define _STRING_H 1
 
+#include <stdint.h>
 #include <stddef.h>
 
+int strlen(const char content[]);
 void *memcpy(void *dest, const void *src, size_t n);
 void *memset(void *s, int c, size_t n);
 void *memmove(void *dest, const void *src, size_t n);
