+++ libc/string/strlen.c
@@ -1,0 +1,7 @@
+#include <include/string.h>
+
+int strlen(char *str) {
+    int len = 0;
+    while (*str++) len++;
+    return len;
+}
