+++ makebackup
@@ -39,12 +39,11 @@ limine/limine:
 	$(MAKE) -C limine
 
 .PHONY: kernel
-kernel: libc
+kernel:
 	$(MAKE) -C kernel
 
-.PHONY: libc
 libc:
-	$(MAKE) -c libc 
+	$(MAKE) -C libc
 
 $(IMAGE_NAME).iso: limine/limine kernel
 	rm -rf iso_root
@@ -81,6 +80,10 @@ $(IMAGE_NAME).hdd: limine/limine kernel
 .PHONY: clean
 clean:
 	rm -rf iso_root $(IMAGE_NAME).iso $(IMAGE_NAME).hdd
-	rm -rf build limine ovmf
+	rm -rf build/
 	$(MAKE) -C kernel clean
-	$(MAKE) -C libc clean
+
+.PHONY: distclean
+distclean: clean
+	rm -rf limine ovmf
+	$(MAKE) -C kernel distclean
